LPS-171473 Embedding inside the operation class
