+++ modules/apps/headless/headless-builder/headless-builder-impl/src/main/java/com/liferay/headless/builder/internal/operation/handler/GetByPrimaryKeyOperationHandler.java
@@ -17,7 +17,6 @@ package com.liferay.headless.builder.internal.operation.handler;
 import com.liferay.headless.builder.internal.constants.HeadlessBuilderConstants;
 import com.liferay.headless.builder.internal.operation.Operation;
 import com.liferay.headless.builder.internal.operation.OperationContext;
-import com.liferay.headless.builder.internal.operation.Response;
 import com.liferay.headless.builder.internal.util.HeadlessBuilderUtil;
 import com.liferay.info.exception.NoSuchInfoItemException;
 import com.liferay.info.field.InfoFieldValue;
@@ -25,6 +24,8 @@ import com.liferay.info.item.provider.InfoItemFieldValuesProvider;
 import com.liferay.info.item.provider.InfoItemObjectProvider;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.vulcan.jaxrs.exception.mapper.Problem;
+
+import javax.ws.rs.core.Response;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -38,7 +39,7 @@ import org.osgi.service.component.annotations.Component;
 public class GetByPrimaryKeyOperationHandler implements OperationHandler {
 
 	@Override
-	public javax.ws.rs.core.Response handle(
+	public Response handle(
 			Operation operation, OperationContext operationContext)
 		throws Exception {
 
@@ -50,9 +51,9 @@ public class GetByPrimaryKeyOperationHandler implements OperationHandler {
 				"The operation context does not contain the primary key");
 		}
 
-		Response response = operation.getResponse(
+		Operation.Response response = operation.getResponse(
 			operationContext.getMediaType(),
-			javax.ws.rs.core.Response.Status.OK.getStatusCode());
+			Response.Status.OK.getStatusCode());
 
 		InfoItemObjectProvider<?> infoItemObjectProvider =
 			HeadlessBuilderUtil.getInfoItemService(
@@ -68,8 +69,8 @@ public class GetByPrimaryKeyOperationHandler implements OperationHandler {
 					response.getEntityName(),
 					InfoItemFieldValuesProvider.class);
 
-			return javax.ws.rs.core.Response.status(
-				javax.ws.rs.core.Response.Status.OK
+			return Response.status(
+				Response.Status.OK
 			).entity(
 				HeadlessBuilderUtil.toDTO(
 					infoItemFieldValuesProvider.getInfoItemFieldValues(object),
@@ -85,10 +86,10 @@ public class GetByPrimaryKeyOperationHandler implements OperationHandler {
 				message = throwable.getMessage();
 			}
 
-			return javax.ws.rs.core.Response.status(
-				javax.ws.rs.core.Response.Status.NOT_FOUND
+			return Response.status(
+				Response.Status.NOT_FOUND
 			).entity(
-				new Problem(javax.ws.rs.core.Response.Status.NOT_FOUND, message)
+				new Problem(Response.Status.NOT_FOUND, message)
 			).build();
 		}
 	}
